requirements.txt
Flask==2.0.1
gunicorn==20.1.0
Werkzeug==2.0.1
openai==0.28.1
python-dotenv==1.0.0
langchain==0.0.200
faiss-cpu==1.7.4
tiktoken==0.5.1
pypdf2